Adapt build and hello_world example to Android NDK
Build adaptations and fixes:
- adapt root CMakeLists.txt to Android NDK build
- add missed port_type_e in internal_android.hpp
- move tp_max_segment_length_ from tp.hpp to tp.cpp due to linker error
- rework message.cpp logging, using <android/log.h> instead of <utils/Log.h> to support both AOSP and NDK builds

examples/hello_world adaptation:
- rework structure to have possibility to build examples like independent executables or
  like header libraries for usage them from client code (e.g. from native part of Android app services)
- update readme_android with NDK usage example
