+++ examples/hello_world/hello_world_client.hpp
@@ -2,11 +2,18 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
-#ifndef VSOMEIP_ENABLE_SIGNAL_HANDLING
-#include <csignal>
-#endif
 #include <vsomeip/vsomeip.hpp>
-#include <iostream>
+
+#ifdef __ANDROID__ // NDK
+#include "android/log.h"
+#define LOG_TAG "hello_world_client"
+#define LOG_INF(...) (void)__android_log_print(ANDROID_LOG_INFO, LOG_TAG, ##__VA_ARGS__)
+#define LOG_ERR(...) (void)__android_log_print(ANDROID_LOG_ERROR, LOG_TAG, ##__VA_ARGS__)
+#else
+#include <cstdio>
+#define LOG_INF(...) fprintf(stdout, __VA_ARGS__)
+#define LOG_ERR(...) fprintf(stderr, __VA_ARGS__)
+#endif
 
 static vsomeip::service_t service_id = 0x1111;
 static vsomeip::instance_t service_instance_id = 0x2222;
@@ -27,7 +34,7 @@ public:
     bool init(){
         // init the application
         if (!app_->init()) {
-            std::cerr << "Couldn't initialize application" << std::endl;
+            LOG_ERR ("Couldn't initialize application");
             return false;
         }
 
@@ -92,7 +99,7 @@ public:
             rq->set_payload(pl);
             // Send the request to the service. Response will be delivered to the
             // registered message handler
-            std::cout << "Sending: " << str << std::endl;
+            LOG_INF("Sending: %s", str.c_str());
             app_->send(rq);
         }
     }
@@ -110,7 +117,7 @@ public:
             std::string resp = std::string(
                     reinterpret_cast<const char*>(pl->get_data()), 0,
                     pl->get_length());
-            std::cout << "Received: " << resp << std::endl;
+            LOG_INF("Received: %s", resp.c_str());
             stop();
         }
     }
@@ -134,31 +141,3 @@ private:
     std::shared_ptr<vsomeip::runtime> rtm_;
     std::shared_ptr<vsomeip::application> app_;
 };
-
-#ifndef VSOMEIP_ENABLE_SIGNAL_HANDLING
-hello_world_client *hw_cl_ptr(nullptr);
-    void handle_signal(int _signal) {
-        if (hw_cl_ptr != nullptr &&
-                (_signal == SIGINT || _signal == SIGTERM))
-            hw_cl_ptr->stop();
-    }
-#endif
-
-int main(int argc, char **argv)
-{
-    (void)argc;
-    (void)argv;
-
-    hello_world_client hw_cl;
-#ifndef VSOMEIP_ENABLE_SIGNAL_HANDLING
-    hw_cl_ptr = &hw_cl;
-    signal(SIGINT, handle_signal);
-    signal(SIGTERM, handle_signal);
-#endif
-    if (hw_cl.init()) {
-        hw_cl.start();
-        return 0;
-    } else {
-        return 1;
-    }
-}
